added check if folder exists in upload form handling (edit)

+++ routes/route.js
@@ -291,12 +291,15 @@ router.post('/upload', function (req, res) {
 });
 
   form.on('end', function(fields, files) {
+    var exists4 = fs.existsSync(invalid_up_loc);
+      if(exists4){
             var testFolder = path.join(invalid_up_loc + '/');
             fs.readdirSync(testFolder).forEach(file=>{
               var filePath = invalid_up_loc + file;
               fs.unlinkSync(filePath);
                 console.log('removed file: ' + file + ' with extension: ' + path.extname(file));
               });
+            }
             console.log("success!");
             res.redirect('/upload_succ');            
         });
